+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Fixed Naive Bayes user-defined class prior probabilities
     - Internal CV Learners now check for sufficient hold out data
     - Fixed randomize empty dataset object
+    - Removed setPersister method from Persistent Model
 
 - 0.1.0-rc5
     - Improved logging for Verbose Learners